requirements.txt
pyspark==3.5.7
kafka-python==2.0.2
requests==2.32.3
PyYAML==6.0.2